backend/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
sqlalchemy==2.0.17
pydantic==2.0.2
alembic==1.11.1
psycopg2-binary==2.9.6
redis==4.6.0
pika==1.3.2
strawberry-graphql==0.194.4
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pydantic-settings==2.0.2
email-validator==2.0.0